refactoring
I couldn't even describe every file that changed, but the basic idea was to refactor and reorganize the whole mission. Everything is a function of some sort now, with the exceptions of external libraries (AIS, COS, etc) and a few legecy files that redirect to their function counterparts (AddBadPoints, etc).